requirements.txt
google-adk>=0.6.0
google-genai>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.10.6